Implemented searching
Implemented searching for employee and purchase. Should be able to
search for text and numbers on an employee, but only text for a
purchase, since i didn’t yet bother fixing the classes or find another
solution for it. Added fields to search in the GUI as well, but the GUI
needs an upgrade in general i think